Fix sign-compare warnings
Mostly due to using signed types and comparing to unsigned types in
for-loops. Change to explicit correct type, or auto if possible.

+++ src/mesh/impls/domain.cxx
@@ -126,7 +126,7 @@ void Domain::addBoundary(Bndry *b) {
 }
 
 void Domain::removeBoundary(Bndry *b) {
-  int s = boundary.size();
+  size_t s = boundary.size();
   boundary.remove(b);
   assert(boundary.size() == s-1);
 }
